// src/types/touch-portal-client-event.ts
export enum TouchPortalClientEvent {
  Connected = 'connected',
  Close = 'Close',
  Info = 'Info',
  Settings = 'Settings',
  NotificationClicked = 'NotificationClicked',
  ListChange = 'ListChange',
  Action = 'Action',
  Broadcast = 'Broadcast',
  ConnectorChange = 'ConnectorChange',
  Message = 'Message',
  ConnectorShortIdNotification = 'ConnectorShortIdNotification',
  Update = 'Update'
}